refactor:componentsフォルダ内にanimationsフォルダを作成、アニメーションに関するコンポーネントを収納
