+++ app/main/page.tsx
@@ -3,8 +3,8 @@
 import React, {useState,useEffect,useRef} from 'react';
 
 // import DogRandomAnimation from '../components/dogRandomAnimation';
-import DogAnimation from '../../components/DogAnimation';
-import CatAnimation from '../../components/CatAnimation';
+import DogAnimation from '../../components/animations/DogAnimation';
+import CatAnimation from '../../components/animations/CatAnimation';
 
 import {getPetInfo} from '../../features/api/getPetInfo'
 import {petPhysicalRecover} from '../../features/api/petPhysicalRecover'
@@ -17,8 +17,8 @@ import { CssBaseline, ThemeProvider, createTheme } from '@mui/material';
 import { CacheProvider } from '@emotion/react';
 import createCache from '@emotion/cache';
 
-import PuppyDogAnimation from '../../components/PuppyDogAnimation';
-import PuppyCatAnimation from '../../components/PuppyCatAnimation';
+import PuppyDogAnimation from '../../components/animations/PuppyDogAnimation';
+import PuppyCatAnimation from '../../components/animations/PuppyCatAnimation';
 
 const cache = createCache({ key: 'css', prepend: true });
