+++ components/animations/CatAnimation.tsx
@@ -2,29 +2,29 @@ import React, { useEffect, useRef, useState, useImperativeHandle, forwardRef } f
 import gsap from 'gsap';
 import Image from 'next/image';
 
-import legImageBackRight from '../public/三毛猫奥後足.png';
-import legImageFrontRight from '../public/三毛猫奥前足.png';
-import legImageFrontLeft from '../public/三毛猫手前前足.png';
-import legImageBackLeft from '../public/三毛猫手間後ろ足.png';
-import tailImage from '../public/三毛猫尻尾.png';
-import faceImage from '../public/三毛猫顔.png';
-import bodyImage from '../public/三毛猫胴体.png';
-import earImage from '../public/三毛猫耳.png';
-import beardImageRight from '../public/三毛猫右ひげ.png';
-import beardImageLeft from '../public/三毛猫左ひげ.png';
+import legImageBackRight from '../../public/三毛猫奥後足.png';
+import legImageFrontRight from '../../public/三毛猫奥前足.png';
+import legImageFrontLeft from '../../public/三毛猫手前前足.png';
+import legImageBackLeft from '../../public/三毛猫手間後ろ足.png';
+import tailImage from '../../public/三毛猫尻尾.png';
+import faceImage from '../../public/三毛猫顔.png';
+import bodyImage from '../../public/三毛猫胴体.png';
+import earImage from '../../public/三毛猫耳.png';
+import beardImageRight from '../../public/三毛猫右ひげ.png';
+import beardImageLeft from '../../public/三毛猫左ひげ.png';
 
-import heartImage from '../public/ハートマーク.png';
-import vesselImage from '../public/容器.png';
-import ballImage from '../public/ボール.png';
+import heartImage from '../../public/ハートマーク.png';
+import vesselImage from '../../public/容器.png';
+import ballImage from '../../public/ボール.png';
 
-import yellowNoteImage from '../public/音符（黄色）.png';
-import blueNoteImage from '../public/音符（青）.png';
+import yellowNoteImage from '../../public/音符（黄色）.png';
+import blueNoteImage from '../../public/音符（青）.png';
 
-import donyoriImage from '../public/どんより1.png';
-import donyori2Image from '../public/どんより2.png';
-import guruguruImage from '../public/ぐるぐる.png';
+import donyoriImage from '../../public/どんより1.png';
+import donyori2Image from '../../public/どんより2.png';
+import guruguruImage from '../../public/ぐるぐる.png';
 
-import SitCatImage from '../public/猫.png';
+import SitCatImage from '../../public/猫.png';
 
 interface CatAnimationHandle {
   feedButtonClick: () => void;
