+++ components/animations/DogAnimation.tsx
@@ -2,28 +2,28 @@ import React, { forwardRef,useEffect, useRef, useState, useImperativeHandle } fr
 import gsap from 'gsap';
 import Image from 'next/image';
 
-import legImageBackRight from '../public/ダックス奥後ろ足.png';
-import legImageFrontRight from '../public/ダックス奥前足2.png';
-import legImageFrontLeft from '../public/ダックス前足.png';
-import legImageBackLeft from '../public/ダックス後ろ足.png';
-import tailImage from '../public/ダックス尻尾.png';
-import headImageFace from '../public/ダックス顔.png';
-import headImageEye from '../public/ダックス目.png';
-import bodyImage from '../public/ダックス胴体.png';
-import earImage from '../public/ダックス耳.png';
-import earImageRight from '../public/ダックス奥耳.png';
-import jawImage from '../public/ダックス顎.png';
+import legImageBackRight from '../../public/ダックス奥後ろ足.png';
+import legImageFrontRight from '../../public/ダックス奥前足2.png';
+import legImageFrontLeft from '../../public/ダックス前足.png';
+import legImageBackLeft from '../../public/ダックス後ろ足.png';
+import tailImage from '../../public/ダックス尻尾.png';
+import headImageFace from '../../public/ダックス顔.png';
+import headImageEye from '../../public/ダックス目.png';
+import bodyImage from '../../public/ダックス胴体.png';
+import earImage from '../../public/ダックス耳.png';
+import earImageRight from '../../public/ダックス奥耳.png';
+import jawImage from '../../public/ダックス顎.png';
 
-import heartImage from '../public/ハートマーク.png';
-import vesselImage from '../public/容器.png';
-import ballImage from '../public/ボール.png';
+import heartImage from '../../public/ハートマーク.png';
+import vesselImage from '../../public/容器.png';
+import ballImage from '../../public/ボール.png';
 
-import yellowNoteImage from '../public/音符（黄色）.png';
-import blueNoteImage from '../public/音符（青）.png';
+import yellowNoteImage from '../../public/音符（黄色）.png';
+import blueNoteImage from '../../public/音符（青）.png';
 
-import donyoriImage from '../public/どんより1.png';
-import donyori2Image from '../public/どんより2.png';
-import guruguruImage from '../public/ぐるぐる.png';
+import donyoriImage from '../../public/どんより1.png';
+import donyori2Image from '../../public/どんより2.png';
+import guruguruImage from '../../public/ぐるぐる.png';
 
 interface DogAnimationHandle {
   feedButtonClick: () => void;
